Simplify shapefile and set projected CRS

Simplifies a shapefile using rmapshaper::ms_simplify() and also sets a projected CRS using sf::st_transform() since rmapshaper::ms_simplify() prefers a projected CRS.

simplify_shapefile(shapefile, projected_crs = 2163, simplify_keep_pct = 0.1)

Arguments

shapefile
An sf object with a raw CoC shapefile from read_raw_coc_shapefile()

projected_crs
The EPSG code for the projected CRS

simplify_keep_pct
The share of the shapefile to keep when simplifying

Value

The simplified sf object now in the projected CRS.

See also